add ticket-detail directive to share in 2 places
