+++ beeline/directives/ticketDetail/ticketDetail.html
@@ -1,0 +1,76 @@
+  <div class="trip-info item item-text-wrap">
+    <div class="row">
+      <div class="col text-center">
+        Ticket for {{ticket.boardStop.time | date:'d MMM y' : '+0800'}}
+      </div>
+    </div>
+    <div class="service-cancellation" ng-if="disp.tripStatus == 'cancelled'">
+      <div class="item item-text-wrap">
+        <div class="cancellation-content">
+          <h2>Service is Cancelled</h2>
+          <img src="img/TicketCancelledTrip.svg" />
+          <!-- <i class="icon ion-alert-circled"></i> -->
+          <p>Today's trip has been cancelled. Sorry for the inconvenience caused. </p>
+          <p>
+            Please make alternative transport arrangements.
+            <strong>{{company.name}}</strong> will contact you and refund today's fare.
+          </p>
+          <p>
+            For enquiries, please contact
+            <strong>{{company.email}}</strong>
+            <strong ng-if="company.contactNo">(Tel: {{company.contactNo}})</strong>
+          </p>
+        </div>
+      </div>
+    </div>
+    <div ng-if="disp.tripStatus !== 'cancelled'">
+      <div class="row">
+        <div class="col col-20">
+          <span class="bus-number" ng-show="route.label">{{route.label}}</span>
+        </div>
+        <div class="col">
+          <h4>Trip Code </h4>
+          <trip-code class="trip-code" code="tripCode" ng-if="tripCode"/>
+          <span ng-if="!tripCode">Not available</span>
+        </div>
+        <div class="col">
+          <h4>Vehicle No.</h4>
+          <div ng-if="disp.vehicle" class="vehicleNo" ng-bind="disp.vehicle">
+          </div>
+          <div ng-if="!disp.vehicle">
+            Not provided
+          </div>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <h4>Operated By</h4>
+          <span ng-bind="company.name"></span>
+          <a ng-click="showTerms(company.id)"><i class="icon ion-information-circled tnc"></i></a>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <h4>Driver</h4>
+          <div ng-bind="disp.driver  || 'Not provided'"></div>
+        </div>
+        <div class="col">
+          <h4>Passenger</h4>
+          <div ng-bind="user.name"></div>
+        </div>
+      </div>
+      <div class="row">
+        <div class="route-info">
+          <div class="start">
+            <span class="time" ng-bind="ticket.boardStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallStart.svg" />
+              <div class="routeDetailDesc">
+                {{ticket.boardStop.stop.description}}, {{ticket.boardStop.stop.road}}</div>
+          </div>
+            <div class="end">
+            <span class="time" ng-bind="ticket.alightStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallEnd.svg" />
+              <div class="routeDetailDesc">{{ticket.alightStop.stop.description}}, {{ticket.alightStop.stop.road}}</div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
