+++ beeline/main.js
@@ -39,6 +39,7 @@ require("./services/RotatedImageService")
 require("./services/GeoUtils")
 require("./directives/mapBusIcon")
 require("./directives/routeItem/animatedRoute")
+require("./directives/ticketDetail/ticketDetail")
 require("./services/fastCheckoutService")
 require("./services/purchaseRoutePassService")
 require("./services/paymentService")
