+++ scss/beeline.scss
@@ -422,39 +422,3 @@ textarea{
   }
 
 }
-
-.trip-info {
-  .route-info{
-    margin: $padding-small-vertical 0;
-    display: flex;
-    flex-direction: column;
-    .start, .end{
-      display: flex;
-      flex-direction: row;
-      align-items: flex-start;
-      padding: $item-padding-small;
-      .routeDetailDesc{
-        flex: 1 1 auto;
-      }
-      .time{
-        white-space: nowrap;
-        flex: 0 0 4em;
-        margin-right: $content-padding;
-      }
-      .route-icon {
-        position: relative;
-        flex: none;
-        vertical-align: top;
-        margin-left: 0;
-        margin-right: $content-padding*0.5;
-        top: $item-padding-small;
-      }
-    }
-  }
-  .bus-number{
-    margin-right: $padding-small-vertical;
-  }
-  .vehicleNo{
-    text-transform: uppercase;
-  }  
-}
