+++ scss/ionic.app.scss
@@ -24,6 +24,7 @@ $ionicons-font-path: "../lib/ionic/fonts";
 @import "scss/progress-bar";
 @import "scss/kickstarterInfo";
 @import "scss/searchButton";
+@import "scss/ticketDetailDirective";
 
 // View specific styles
 @import "scss/tabs";
