+++ scss/ticketDetail.scss
@@ -8,105 +8,4 @@
     top: auto;
     min-height: 200px;
   }
-  .item.trip-info {
-    flex: 0 0 auto;
-    background-color: rgba($light, 0.9);
-    padding: $item-padding-small;
-    h4 {
-      color: $normal-grey;
-    }
-    .tnc {
-      padding: 0 $item-padding-base;
-    }
-    .ion-information-circled {
-      color: $royal;
-      font-size: 1.5em;
-    }
-    .route-info{
-      margin: $padding-small-vertical 0;
-      display: flex;
-      flex-direction: column;
-      .start, .end{
-        display: flex;
-        flex-direction: row;
-        align-items: flex-start;
-        padding: $item-padding-small;
-        .routeDetailDesc{
-          flex: 1 1 auto;
-        }
-        .time{
-          white-space: nowrap;
-          flex: 0 0 4em;
-          margin-right: $content-padding;
-        }
-        .route-icon {
-          position: relative;
-          flex: none;
-          vertical-align: top;
-          margin-left: 0;
-          margin-right: $content-padding*0.5;
-          top: $item-padding-small;
-        }
-      }
-    }
-    .bus-number{
-      margin-right: $padding-small-vertical;
-    }
-    .vehicleNo{
-      text-transform: uppercase;
-    }
-  }
-  .service-cancellation {
-    flex: 1 1 auto;
-    .cancellation-content {
-      margin: $padding-large-vertical 0;
-      text-align: center;
-      h2 {
-        color: $normal-grey;
-        text-align: center;
-        font-weight: bold;
-        padding: $item-padding-base 0;
-      }
-      img {
-        padding: $line-height-computed 0;
-      }
-      p {
-        text-align: left;
-        padding: $padding-base-vertical $padding-base-horizontal;
-      }
-    }
-  }
-
-  .tab-item {
-    &.active {
-      .ticket-nav-ticket,
-      .ticket-nav-moreInfo {
-        opacity: 1;
-      }
-    }
-    &:not(.active) {
-      .ticket-nav-ticket,
-      .ticket-nav-moreInfo {
-        opacity: 0.3;
-      }
-    }
-  }
-
-  .ticket-nav-ticket {
-    background-image: url('../img/ticketdetail_ticket_normal.svg');
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    margin-right: $padding_small_horizontal*0.5;
-    vertical-align: sub;
-  }
-
-  .ticket-nav-moreInfo {
-    background-image: url('../img/ticketdetail_moreinfo_normal.svg');
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    margin-right: $padding_small_horizontal*0.5;
-    vertical-align: sub;
-  }
 }
