+++ scss/ticketDetailDirective.scss
@@ -1,0 +1,56 @@
+.trip-info {
+  .route-info{
+    margin: $padding-small-vertical 0;
+    display: flex;
+    flex-direction: column;
+    .start, .end{
+      display: flex;
+      flex-direction: row;
+      align-items: flex-start;
+      padding: $item-padding-small;
+      .routeDetailDesc{
+        flex: 1 1 auto;
+      }
+      .time{
+        white-space: nowrap;
+        flex: 0 0 4em;
+        margin-right: $content-padding;
+      }
+      .route-icon {
+        position: relative;
+        flex: none;
+        vertical-align: top;
+        margin-left: 0;
+        margin-right: $content-padding*0.5;
+        top: $item-padding-small;
+      }
+    }
+  }
+  .bus-number{
+    margin-right: $padding-small-vertical;
+  }
+  .vehicleNo{
+    text-transform: uppercase;
+  }
+}
+
+.service-cancellation {
+  flex: 1 1 auto;
+  .cancellation-content {
+    margin: $padding-large-vertical 0;
+    text-align: center;
+    h2 {
+      color: $normal-grey;
+      text-align: center;
+      font-weight: bold;
+      padding: $item-padding-base 0;
+    }
+    img {
+      padding: $line-height-computed 0;
+    }
+    p {
+      text-align: left;
+      padding: $padding-base-vertical $padding-base-horizontal;
+    }
+  }
+}
