+++ static/templates/kickstarter-detail.html
@@ -1,8 +1,8 @@
 <ion-view id="pg_kickstarter_detail" view-title="Crowdstart Route Details"
- can-swipe-back="false" class="kickstarter-detail2 page-with-map">
+ can-swipe-back="false" class="kickstarter-detail2 page-with-map" hide-back-button="true">
  <ion-nav-buttons side="left">
    <button class="button button-clear back-button"
-      ng-click="goBack()">
+      ui-sref="tabs.crowdstart">
       <i class="icon ion-ios-arrow-back"></i>
    </button>
  </ion-nav-buttons>
