+++ static/templates/route-detail.html
@@ -75,42 +75,13 @@
       </div>
     </div>
     <div ng-show="activeTab===1" class="ticketView">
-      <div class="booking-info trip-info">
-        <div class="row">
-          <div class="col text-center">
-            {{disp.navTitle}}
-          </div>
-          <div class="col col-20 text-right">
-            <i class="fa fa-times fa-2x" aria-hidden="true" ng-click="toggle()"></i>
-          </div>
+      <div class="row">
+        <!-- delay directive rendering until nextTripTicketId is ready -->
+        <div class="col" ng-if="data.nextTripTicketId">
+          <ticket-detail ticket-id="data.nextTripTicketId"></ticket-detail>
         </div>
-        <div class="row">
-          <div class="col">
-            <h4>Trip Code </h4>
-            <trip-code class="trip-code" code="disp.tripCode" ng-if="disp.tripCode">
-            </trip-code>
-            <span ng-if="!disp.tripCode">Not available</span>
-          </div>
-          <div class="col">
-            <h4>Vehicle No.</h4>
-            <div ng-if="disp.vehicle" class="vehicleNo" ng-bind="disp.vehicle">
-            </div>
-            <div ng-if="!disp.vehicle">
-              Not provided
-            </div>
-          </div>
-        </div>
-        <div class="row">
-          <div class="route-info">
-            <div class="start">
-              <span class="time" ng-bind="ticket.boardStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallStart.svg" />
-              <div class="routeDetailDesc">{{ticket.boardStop.stop.description}}, {{ticket.boardStop.stop.road}}</div>
-            </div>
-            <div class="end">
-              <span class="time" ng-bind="ticket.alightStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallEnd.svg" />
-              <div class="routeDetailDesc">{{ticket.alightStop.stop.description}}, {{ticket.alightStop.stop.road}}</div>
-            </div>
-          </div>
+        <div>
+          <i class="fa fa-times fa-lg" aria-hidden="true" ng-click="toggle()"></i>
         </div>
       </div>
    </div>
